+++ examples/choice.py
@@ -7,8 +7,14 @@ from pydantic import BaseModel
 class MyModel(BaseModel):
     """My model with choice values."""
 
-    my_value: Literal["option1", "option2"] = "option1"
-    """A value with restricted values."""
+    required_choice: Literal["option1", "option2"]
+    """Required restricted values."""
+
+    optional_choice: Literal[1, 2] = 2
+    """Optional with a literal default."""
+
+    optional_with_none: Literal[4, 5] | None = None
+    """Optional with None as default."""
 
 
-model = parse(MyModel)
+print(repr(parse(MyModel)))