+++ examples/keyword.py
@@ -1,12 +1,20 @@
-from clipstick import parse
+from typing import Annotated
+
+from clipstick import parse, short
 from pydantic import BaseModel
 
 
 class MyModel(BaseModel):
-    """A model with a keyworded optional value."""
+    """A model with keyworded optional values."""
 
     my_value: int = 22
     """My value with a default."""
 
+    other_value: int | None = None
+    """Value with None as default."""
 
-print(parse(MyModel))
+    with_short: Annotated[str, short("w")] = "some_value"
+    """With a shorthand key."""
+
+
+print(repr(parse(MyModel)))