+++ CHANGELOG.md
@@ -9,6 +9,15 @@ and this project adheres to [Semantic Versioning].
 
 - /
 
+## [0.5.0] - 2024-01-08
+
+### Added
+
+- None as a default value. See overal Usage documentation.
+- Add python 3.12 to testing matrix
+
+
+
 ## [0.4.2] - 2024-01-04
 
 ### Fixed